proper paths are everywhere and all user actions are accessed by screen

+++ cells_dir/kill_cell.cpp
@@ -46,7 +46,7 @@ type kill_cell::getCellType() const {
 }
 
 std::ostream &operator<<(std::ostream &out, const kill_cell &ref) {
-    out<<ref.getCellType();
+    out<<(int)ref.getCellType();
     return out;
 }
